✨ feat: 📓 Add new notebook and 📝 update existing filenames
- 🚀 Added a fresh notebook to enhance our documentation.
- 🔎 Updated filenames for better clarity and organization.
