+++ custom_diffusion/automodel.py
@@ -1,54 +1,34 @@
-# Mapping between task identifiers and their respective class names
+from custom_diffusion.pipelines import (
+    StableDiffusionControlNetGenerator,
+    StableDiffusionControlNetImg2ImgGenerator,
+    StableDiffusionControlNetInpaintGenerator,
+    StableDiffusionImg2ImgGenerator,
+    StableDiffusionText2ImgGenerator,
+    StableDiffusionUpscaleGenerator,
+)
+
 TASK_ID_TO_CLASS_MAPPING = {
-    "stable-txt2img": "StableDiffusionText2ImgGenerator",
-    "stable-img2img": "StableDiffusionImg2ImgGenerator",
-    "stable-inpaint": "StableDiffusionInpaintingGenerator",
-    "stable-upscale": "StableDiffusionUpscaleGenerator",
-    "stable-pix2pix": "StableDiffusionInstructPix2PixGenerator",
-    "controlnet": "StableDiffusionControlNetGenerator",
-    "controlnet-img2img": "StableDiffusionControlNetImg2ImgGenerator",
-    "controlnet-inpaint": "StableDiffusionControlNetInpaintGenerator",
+    "stable-txt2img": StableDiffusionText2ImgGenerator,
+    "stable-img2img": StableDiffusionImg2ImgGenerator,
+    "stable-upscale": StableDiffusionUpscaleGenerator,
+    "controlnet": StableDiffusionControlNetGenerator,
+    "controlnet-img2img": StableDiffusionControlNetImg2ImgGenerator,
+    "controlnet-inpaint": StableDiffusionControlNetInpaintGenerator,
 }
 
 
-class AutoDiffusionModel:
-    """
-    Factory class for automatically generating diffusion models based on task IDs.
-    Provides a convenient interface to load models without knowing the underlying classes.
-    """
+def pipeline(task_id: str, model_id: str, scheduler_name: str = "DDIM"):
+    # TASK_ID_TO_CLASS_MAPPING dictionary is assumed to be defined as before
 
-    @staticmethod
-    def from_pretrained(
-        task_id: str = "stable-txt2img",
-        model_path: str = "runwayml/stable-diffusion-v1-5",
-        scheduler_name: str = "DDIM",
-        **kwargs,
-    ):
-        """
-        Load a diffusion model based on its task ID, model path, scheduler name, and other parameters.
+    # Get the class directly from the mapping based on task_id
+    DiffusionModelClass = TASK_ID_TO_CLASS_MAPPING.get(task_id)
+    if not DiffusionModelClass:
+        raise ValueError(f"Unsupported task ID: {task_id}")
 
-        Args:
-        - task_id (str, optional): Identifier for the type of diffusion task. Defaults to 'stable-txt2img'.
-        - model_path (str, optional): Path to the pretrained weights. If not provided, model is initialized with default path.
-        - scheduler_name (str, optional): Name of the scheduler. Defaults to 'DDIM'.
-        - **kwargs: Additional arguments required by the specific diffusion model.
+    # Instantiate the model class
+    model_instance = DiffusionModelClass()
 
-        Returns:
-        - model_instance (BaseDiffusionModel): Initialized diffusion model.
-        """
+    # Load the model based on model_id
+    model_instance.load_model(model_path=model_id, scheduler_name=scheduler_name)
 
-        # Get the class name based on task ID
-        model_class_name = TASK_ID_TO_CLASS_MAPPING.get(task_id)
-        if not model_class_name:
-            raise ValueError(f"Unsupported task ID: {task_id}")
-
-        # Dynamically import the required class using its name and the task ID
-        DiffusionModelClass = getattr(
-            __import__(f"custom_diffusion.pipelines.{task_id}", fromlist=[model_class_name]), model_class_name
-        )
-
-        # Return an instance of the loaded model class, initialized with the provided parameters
-        model_instance = DiffusionModelClass()
-        model_instance.load_model(model_path=model_path, scheduler_name=scheduler_name, **kwargs)
-
-        return model_instance
+    return model_instance
