+++ custom_diffusion/pipelines/__init__.py
@@ -1,0 +1,6 @@
+from .controlnet import StableDiffusionControlNetGenerator
+from .controlnet_img2img import StableDiffusionControlNetImg2ImgGenerator
+from .controlnet_inpaint import StableDiffusionControlNetInpaintGenerator
+from .stable_img2img import StableDiffusionImg2ImgGenerator
+from .stable_txt2img import StableDiffusionText2ImgGenerator
+from .stable_upscale import StableDiffusionUpscaleGenerator
